bug修正

+++ enhance/src/main/java/com/xzixi/self/portal/enhance/annotation/processor/CacheEnhanceProcessor.java
@@ -266,7 +266,7 @@ public class CacheEnhanceProcessor extends AbstractBaseProcessor {
         // 方法名
         Name name = getNameFromString(method);
         // 返回值类型
-        JCTree.JCExpression returnType = treeMaker.TypeIdent(TypeTag.INT);
+        JCTree.JCExpression returnType = treeMaker.TypeIdent(TypeTag.LONG);
         // 参数列表
         List<JCTree.JCVariableDecl> parameters = List.of(queryParamsParamDecl(queryParams, modelClassName));
         // 方法体
